Felefold fixes (#854)
- Prefer the already set sh.userid and sh.groupid over outright calls to
  getuid() and getgid().
- Backport minor lint fixes from ksh2020 for redundant else
  statements (commits 4116459 and 18823be).
- Add include guards to headers that were missing them.
- xec.c: After forking a child process, use sh.current_pid directly
  rather than misleadingly using 'parent' as the child process PID.
- xec.c: Remove unused argument for the unset_instance() function.
- Add the noreturn attribute to sh_main.
- kshrc.sh: Use a red color when the exit status is a negative number
  (e.g. a shell function returned -1).
- optget: Remove obsolete compat code for UWIN.
- tee: Remove unused code for ancient BSD.
- options.sh: Fix tests on Haiku again to cope for Haiku fixing their
  implementation of SIGKILL (which now properly outputs 'Killed').
- terminal.h: Removed extern for the no longer existing sh_ioctl().
- Patch iffe to prevent it from leaking $PACKAGEROOT while making
  the ast/dlldefs.h file; i.e., remove it from this line (re: 9b64328):
  #define _LIB_ast	1	/* /home/johno/GitRepos/KornShell/ksh/arch/linux.i386-64/lib/libast.a is a library */
- mamake: For efficiency, check MAMAKE_STRICT before calling strncmp.
- Avoid passing NULL to gettext to fix a GCC analyzer warning.
- Removed null pointer arithmetic in the grep builtin.
- Moved the AST_LC_* bitflags into the 32-bit range a second time
  (re: 0a2302e2).
- Fixed a -Wsign-compare warning in utf8_mbtowc() by changing the
  int `m' to size_t. This seemingly innocuous change results in a
  small speedup in shellbench for ksh binaries on Linux x86_64
  (10 runs of charsplit with CCFLAGS=-O2):
     -------------------------------------------------------
     name           /tmp/ksh-int-gcc     /tmp/ksh-size_t-gcc
     -------------------------------------------------------
     charsplit.ksh  0.425 [0.422-0.431]  0.325 [0.289-0.358]
     -------------------------------------------------------
     ---------------------------------------------------------
     name           /tmp/ksh-int-clang   /tmp/ksh-size_t-clang
     ---------------------------------------------------------
     charsplit.ksh  0.391 [0.389-0.393]  0.357 [0.355-0.360]
     ---------------------------------------------------------

+++ src/lib/libast/comp/setlocale.c
@@ -508,7 +508,7 @@ static int
 utf8_mbtowc(wchar_t* wp, const char* str, size_t n)
 {
 	unsigned char*	sp = (unsigned char*)str;
-	int		m;
+	size_t		m;
 	int		i;
 	int		c;
 	wchar_t		w = 0;
